+++ lib/maru/parsers/json.ex
@@ -1,0 +1,45 @@
+defmodule Maru.Parsers.JSON do
+  @moduledoc """
+  Fork from Plug.Parsers.JSON, version: plug v1.0.2
+  """
+
+  @behaviour Plug.Parsers
+  import Plug.Conn
+
+  def parse(conn, "application", subtype, _headers, opts) do
+    if subtype == "json" || String.ends_with?(subtype, "+json") do
+      decoder = Keyword.get(opts, :json_decoder) ||
+                  raise ArgumentError, "JSON parser expects a :json_decoder option"
+      case conn.private.maru_body do
+        nil -> read_body(conn, opts)
+        {:ok, body} -> {:ok, body, conn}
+        {:more, body} -> {:more, body, conn}
+      end |> decode(decoder)
+    else
+      {:next, conn}
+    end
+  end
+
+  def parse(conn, _type, _subtype, _headers, _opts) do
+    {:next, conn}
+  end
+
+  defp decode({:more, _, conn}, _decoder) do
+    {:error, :too_large, conn}
+  end
+
+  defp decode({:ok, "", conn}, _decoder) do
+    {:ok, %{}, conn}
+  end
+
+  defp decode({:ok, body, conn}, decoder) do
+    case decoder.decode!(body) do
+      terms when is_map(terms) ->
+        {:ok, terms, conn}
+      terms ->
+        {:ok, %{"_json" => terms}, conn}
+    end
+  rescue
+    e -> raise Plug.Parsers.ParseError, exception: e
+  end
+end