+++ lib/maru/plugs/prepare.ex
@@ -10,7 +10,7 @@ defmodule Maru.Plugs.Prepare do
 
   @doc false
   def call(conn, []) do
-    conn |> check_path |> check_route_path |> check_param_context |> check_version
+    conn |> check_path |> check_route_path |> check_param_context |> check_version |> check_body
   end
 
   defp check_path(%Conn{private: %{maru_resource_path: _}}=conn), do: conn
@@ -24,4 +24,7 @@ defmodule Maru.Plugs.Prepare do
 
   defp check_version(%Conn{private: %{maru_version: _}}=conn), do: conn
   defp check_version(conn), do: Conn.put_private(conn, :maru_version, nil)
+
+  defp check_body(%Conn{private: %{maru_body: _}}=conn), do: conn
+  defp check_body(conn), do: Conn.put_private(conn, :maru_body, nil)
 end